Rewrite almost everything
Making room for the new parser. The new lexer is now in place and seems to work at the moment, despite missing the ability to implement loops. Soon... :^)
