+++ source/Heavy.py
@@ -1,9 +1,37 @@
 import random
-from Stack import Stack
+class Stack: 
+    def __init__(self):
+        self.items = []
+
+    def is_empty(self):
+        return self.items == []
+
+    def push(self, item:int):
+        self.items.insert(0, item)
+        
+    def pop(self):
+        if len(self.items) == 0:
+            return 
+        else:
+            self.items.pop(0)
+
+    def print_stack(self):
+        print(self.items)
+        
+    def get_first(self):
+        if len(self.items) == 0:
+            return 0
+        else:
+            return self.items[0]
+
+    def get_all_items(self):
+        return self.items
+    
+    def get_length(self):
+        return len(self.items)
 
 stack = Stack()
-
-class Colours:
+class Commands:
     acc1 = 0
     acc2 = 0
     def __init__(self):
@@ -67,4 +95,4 @@ class Colours:
         self.acc2 = 0
 
     def print_stack(self):
-        print(stack.items)
+        return stack.items
