+++ source/_filename.sus
@@ -1,27 +1,66 @@
-ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
-ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
-ORANGE SUS RED SUS RED SUS RED SUS RED SUS RED SUS
+ORANGE SUS
 BLUE SUS
 
+ORANGE SUS ORANGE SUS
+RED SUS RED SUS RED SUS
+BLUE SUS
+
+WHITE SUS WHITE SUS WHITE SUS
+ORANGE SUS ORANGE SUS ORANGE SUS
+ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
+BLUE SUS
+
+ORANGE SUS
+WHITE SUS WHITE SUS
+BLUE SUS
+
+RED SUS RED SUS RED SUS
+RED SUS RED SUS RED SUS
+BLUE SUS
+
+WHITE SUS WHITE SUS WHITE SUS
+BLUE SUS
+
+ORANGE SUS
+WHITE SUS WHITE SUS
+BLUE SUS
+
+PINK SUS
+ORANGE SUS ORANGE SUS ORANGE SUS
+RED SUS RED SUS
+BLUE SUS
+
+ORANGE SUS
+RED SUS RED SUS
+BLUE SUS
+
+ORANGE SUS ORANGE SUS ORANGE SUS
+ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
+WHITE SUS WHITE SUS WHITE SUS
+BLUE SUS
+
+WHITE SUS WHITE SUS WHITE SUS
+BLUE SUS
+BLUE SUS
+
+WHITE SUS WHITE SUS WHITE SUS
+WHITE SUS WHITE SUS WHITE SUS WHITE SUS
+BLUE SUS
+
+PINK SUS
+ORANGE SUS ORANGE SUS ORANGE SUS
+ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
 RED SUS RED SUS
 BLUE SUS
 
 PINK SUS
-ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS ORANGE SUS
-ORANGE SUS ORANGE SUS ORANGE SUS
-RED SUS RED SUS RED SUS
+VENTED VENTED
+SUSSY SUSSY SUSSY
+SUSSY SUSSY SUSSY
 BLUE SUS
-
-PINK SUS
-BLUE SUS 
-
-ELECTRICAL
-VENTED 
-SUSSY SUSSY SUSSY SUSSY SUSSY SUSSY
-
 WHO?
 PURPLE SUS
 GREEN SUS
 PURPLE SUS
 SUSSY BLUE SUS
-WHERE? 
+WHERE?
